% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.dampen.R
\name{fit.dampen}
\alias{fit.dampen}
\title{Runs the model fitting}
\usage{
fit.dampen(y, model = c("Dampen", "SDampen", "TDampen", "TSDampen"),
  control = lgt.control(), nChains = 2, nCores = 2, addJitter = TRUE,
  verbose = FALSE)
}
\arguments{
\item{y}{the time series}

\item{model}{a stan model}

\item{control}{control arguments list}

\item{nChains}{number of MCMC chains . Must >=1. Perhaps optimal number is 4.}

\item{nCores}{number of cores to be used. For performance reasons it should be equal to nChains, 
but nChains should be smaller or equal to the number of cores on the computer.}

\item{addJitter}{adding a bit of jitter is helping Stan in case of some flat series}

\item{verbose}{print verbose information yes/no}
}
\value{
dampenmodel
}
\description{
Runs the model fitting
}
\examples{
\dontrun{
lgt_model <- fit.lgt(lynx, model="LGT", nCores=4, nChains=4,
control=lgt.control(MAX_NUM_OF_REPEATS=1, NUM_OF_ITER=2000), 
verbose=TRUE)
# print the model details
print(lgt_model)
}

\dontrun{demo(exampleScript)}

}
