% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{fit.lgt}
\alias{fit.lgt}
\title{Runs the model fitting}
\usage{
fit.lgt(y, stanModel = NULL, control = lgt.control(), ncores = 1,
  addJitter = TRUE, verbose = FALSE)
}
\arguments{
\item{y}{the time series}

\item{stanModel}{a stan model}

\item{control}{control arguments list}

\item{ncores}{number of cores}

\item{addJitter}{adding a bit of jitter is helping Stan in case of some flat series}

\item{verbose}{print verbose information yes/no}

\item{h}{prediction horizon}
}
\value{

}
\description{
Runs the model fitting
}
