% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{forecast.lgt}
\alias{forecast.lgt}
\title{produce forecasts}
\usage{
\method{forecast}{lgt}(object, h = ifelse(frequency(object$x) > 1, 2 *
  frequency(object$x), 10), NUM_OF_TRIALS = 5000, MIN_VAL = 0.001,
  MAX_VAL = 1e+38, ...)
}
\arguments{
\item{object}{}

\item{h}{}

\item{NUM_OF_TRIALS}{}

\item{MIN_VAL}{}

\item{MAX_VAL}{}

\item{...}{}
}
\value{
returns a forecast object compatible with the forecast package
}
\description{
This function produces forecasts from a model
}
\author{
bergmeir
}
