% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{init.lgt}
\alias{init.lgt}
\title{Initialize a non-seasonal LGT stan model}
\usage{
init.lgt(silent = TRUE)
}
\arguments{
\item{silent}{output stan compiler messages yes/no}
}
\value{

}
\description{
Initialize a stan model that uses the (non-seasonal) LGT
}
