% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{init.trend}
\alias{init.trend}
\title{Initialize a stan model that uses trend only}
\usage{
init.trend(silent = TRUE)
}
\arguments{
\item{silent}{output stan compiler messages yes/no}
}
\value{

}
\description{
Initialize a stan model that is a trend-only, Gaussian error, homoscedastic version of LGT
}
