% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R, R/lgt.R
\name{initModel}
\alias{initModel}
\alias{initModel}
\title{Initialize a non-seasonal LGT stan model}
\usage{
initModel(modelType = NULL)

initModel(modelType = NULL)
}
\arguments{
\item{modelType}{type of the forecasting model selected}

\item{modelType}{type of the forecasting model selected}
}
\value{
SkeletonModel

SkeletonModel
}
\description{
Initialize a stan model that uses the (non-seasonal) LGT

Initialize a stan model that uses the (non-seasonal) LGT
}
\details{
This function is only used as a building block for the main functions

This function is only used as a building block for the main functions
}
