% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgt.R
\name{lgt.control}
\alias{lgt.control}
\title{Sets and initializes the main parameters of the algorithm}
\usage{
lgt.control(MAX_RHAT_ALLOWED = 1.005, NUM_OF_ITER = 5000,
  MAX_NUM_OF_REPEATS = 3, CAUCHY_SD_DIV = 200, MIN_SIGMA = 0.001,
  MIN_NU = 1.5, MAX_NU = 20, MIN_POW = -0.5, MAX_POW = 1)
}
\arguments{
\item{MAX_RHAT_ALLOWED}{}

\item{NUM_OF_ITER}{}

\item{MAX_NUM_OF_REPEATS}{}

\item{CAUCHY_SD_DIV}{}

\item{MIN_SIGMA}{}

\item{MIN_NU}{}

\item{MAX_NU}{}

\item{MIN_POW}{}

\item{MAX_POW}{}
}
\description{
This is a function that initializes and sets the parameters of the algorithm. 
It generates a list of parameters, to be used with the \code{\link{fit.lgt}} function.
}
