% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rlgtfit.R
\name{print.rlgtfit}
\alias{print.rlgtfit}
\alias{summary.rlgt}
\title{Generic print function for rlgtfit models}
\usage{
\method{print}{rlgtfit}(x, ...)
}
\arguments{
\item{x}{an rlgtfit object}

\item{...}{additional function parameters (currently not used)}
}
\description{
Print out some characteristics of an \code{\link{rlgtfit}} model.
}
