% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.dampen.R
\name{posterior_interval.dampen}
\alias{posterior_interval.dampen}
\title{dampen posterior interval}
\usage{
\method{posterior_interval}{dampen}(object, prob = 0.9, type = "central",
  ...)
}
\arguments{
\item{object}{an object of class dampen}

\item{prob}{percentile level to be generated (multiple values can be accepted as a vector)}

\item{type}{currently only central is available}

\item{...}{currently not in use}
}
\value{
confidence interval
}
\description{
This is a method of dampen object to produce posterior interval
}
\examples{
\dontrun{
dampen_model <- fit.dampen(lynx, model="dampen", nCores=4, nChains=4,
control=dampen.control(MAX_NUM_OF_REPEATS=1, NUM_OF_ITER=2000), 
verbose=TRUE)
# print the model details
posterior_interval(dampen_model)
}
}
\author{
wibowo
}
