% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval.lgt.R
\name{posterior_interval.lgt}
\alias{posterior_interval.lgt}
\title{lgt posterior interval}
\usage{
\method{posterior_interval}{lgt}(object, prob = 0.9, type = "central", ...)
}
\arguments{
\item{object}{an object of class lgt}

\item{prob}{percentile level to be generated (multiple values can be accepted as a vector)}

\item{type}{currently only central is available}

\item{...}{currently not in use}
}
\value{
confidence interval
}
\description{
This is a method of lgt object to produce posterior interval
}
\examples{
\dontrun{
lgt_model <- fit.lgt(lynx, model="LGT", nCores=4, nChains=4,
control=lgt.control(MAX_NUM_OF_REPEATS=1, NUM_OF_ITER=2000), 
verbose=TRUE)
# print the model details
posterior_interval(lgt_model)
}
}
\author{
wibowo
}
