% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rlgt-package.R
\docType{package}
\name{Rlgt-package}
\alias{Rlgt-package}
\alias{Rlgt}
\title{Getting started with the Rlgt package}
\examples{
x <- 1
}
\author{
Slawek Smyl \email{slaweks@hotmail.co.uk} 

Christoph Bergmeir \email{christoph.bergmeir@gmail.com}
}
\keyword{exponential}
\keyword{forecasting,}
\keyword{smoothing}
