% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{initModel}
\alias{initModel}
\title{Initialize a model from the Rlgt family}
\usage{
initModel(model.type = NULL, use.regression = FALSE,
  seasonalityMethodId = 0, levelMethodId = 0,
  useSmoothingMethodForError = FALSE)
}
\arguments{
\item{model.type}{type of the forecasting model selected, a character object}

\item{use.regression}{binary parameter indicating whether additional regressors will be used for forecasting in multivariate settings.}

\item{seasonalityMethodId}{Seasonality method Id (0- HW, 1- generalized).}

\item{levelMethodId}{Level method Id.}

\item{useSmoothingMethodForError}{if the non-standard function for error size should be used, one based on smoothed innovations or surprises}
}
\value{
an Rlgt skeleton model
}
\description{
This is an internal function that usually won't be called by users directly. It validates the model type and generates the corresponding list of parameters for the model.
}
